+++ source/src/dict/Ostap.hh
@@ -110,6 +110,7 @@
 #include "Ostap/MoreMath.h"
 #include "Ostap/MoreRooFit.h"
 #include "Ostap/MoreVars.h"
+#include "Ostap/MtrxExprs.h"
 #include "Ostap/Mute.h"
 #include "Ostap/Norms.h"
 #include "Ostap/NSphere.h"