+++ ReleaseNotes/release_notes.md
@@ -18,6 +18,8 @@
    1. add `VIF` methods to get the variance inflation factors for badly-defined covariance matrices 
    1. re-factor some function from `Ostap/Math.h` allowing the separate access to them without heavy `Math.h`
    1. complete linalg-tests
+   1. add "transpose" constructor to GSL =matrix
+   1. add "diagonal matrix expressions" for SMatrix
    
 ## Bug fixes